+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - current page can be pinned
 - added :tabdo to run vi commands on each tab
 - added recursive option
+- script to upload files to sourceforge
 
 ### Changed